Dto inicial da Obra

+++ Livros.Domain/Entities/Obra.cs
@@ -11,7 +11,7 @@ namespace Livros.Domain.Entities
         public string Anotacao { get; set; }
         public int AvaliacaoTotal { get; set; }
         public decimal PrecoTotal { get; set; }
-        public decimal PaginaTotal { get; set; }
+        public int PaginaTotal { get; set; }
         public bool Concluido { get; set; }
         public bool VolumeUnico { get; set; }
         public string Tipo { get; set; }
